Migrate to sbt 0.13.16
